api/requirements.txt
Django==4.1.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3
requests==2.28.1
